Restored the correct return type of the jsonToObject method. Moved the jsonToObject method from the MoleculeStructure class to the Molecule class to avoid cyclic import references
