+++ src/Molecule.js
@@ -833,4 +833,47 @@ export default class Molecule extends MoleculeStructure {
   generateIndex () {
     return Molecule.generateNextAtomIndex( this.atoms );
   }
+
+  /**
+   * Converts a JSON object into a Molecule Structure instance
+   *
+   * @param {string} json
+   * @return {object}
+   * @throws {AtomsMissingException}
+   */
+  static jsonToObject ( json ) {
+    const target = Object.assign( new Molecule( {} ), JSON.parse( json ) ),
+      properties = Object.keys( new Molecule( {} ) );
+
+    if ( !Array.isArray( target.atoms ) ) {
+      throw new AtomsMissingException();
+    }
+
+    for ( const index in Object.keys( target.atoms ) ) {
+
+      target.atoms[ index ] = Atom.jsonToObject( JSON.stringify( target.atoms[ index ] ) );
+
+      for ( const property of [ 'position', 'walletAddress', 'isotope' ] ) {
+
+        if ( typeof target.atoms[ index ][ property ] === 'undefined'
+          || null === target.atoms[ index ][ property ]
+        ) {
+          throw new AtomsMissingException( 'MolecularStructure::jsonToObject() - Required Atom properties are missing!' );
+        }
+      }
+    }
+
+    for ( const property in target ) {
+
+      if ( target.hasOwnProperty( property )
+        && !properties.includes( property )
+      ) {
+        delete target[ property ];
+      }
+    }
+
+    target.atoms = Atom.sortAtoms( target.atoms );
+
+    return target;
+  }
 }
